Use an Angular Material dialog for piece deletion confirmation

+++ src/app/library/piece/piece.component.ts
@@ -1,8 +1,11 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import {MatDialog } from '@angular/material/dialog';
+
 import { LibraryService } from '../library.service';
 import { PieceType } from '../model/piece-type.enum';
 import { Piece } from '../model/piece.model';
+import { PieceDeletionDialogComponent } from './piece-deletion-dialog/piece-deletion-dialog.component';
 
 
 @Component({
@@ -16,7 +19,8 @@ export class PieceComponent implements OnInit {
   panelOpenState: boolean;
 
   constructor(private libraryService: LibraryService,
-              private router: Router) { }
+              private router: Router,
+              public dialog: MatDialog) { }
 
   ngOnInit(): void {
   }
@@ -53,15 +57,26 @@ export class PieceComponent implements OnInit {
 
 
     // TODO add a confirmation popup (potentially with Angular Material MatDialog component)
+    // https://material.angular.io/components/dialog/overview
 
     onDelete(event: Event): void {
     // prevent the card to open when the Delete button is clicked
     event.stopPropagation();
-    this.libraryService.deletePiece(this.piece.id).subscribe(
-      (_) => {
-        this.libraryService.fetchPieces(null, null);
-        this.router.navigate(['library']);
+
+    // Confirmation popup with Angular Material : https://material.angular.io/components/dialog/overview
+    const dialogRef = this.dialog.open(PieceDeletionDialogComponent, {
+      width: '500px',
+      data: {pieceTitle: this.piece.title}
+    });
+    dialogRef.afterClosed().subscribe(confirmDeletion => {
+      if (confirmDeletion) {
+        this.libraryService.deletePiece(this.piece.id).subscribe(
+          (_) => {
+            this.libraryService.fetchPieces(null, null);
+            this.router.navigate(['library']);
+          }
+        );
       }
-    );
+    });
   }
 }
